FunctionsJSON move to cpp file
